Added support for -port parameter, which allows you to specify one or more port name matches (regular expressions). The original behavior is preserved by making the default "system:(capture|monitor)_"
Added support for an -offset parameter, which allows you to skip the
first N ports that match. This is helpful if you have a lot of channels
and want to iterate over them with the command line, for example to
find where a buzzing microphone is coming from.

Added support for a -names parameter, which similar to -index will
display the name of each port. This makes it a lot easier to identify
what audio you are monitoring.

Added support for a -verbose parameter, which currently just displays
a message indicating which ports it is monitoring.